pyproject.toml
[project]
name = "movietracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Carine Neris"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.110.0)",
    "uvicorn[standard] (>=0.29.0)",
    "sqlalchemy (>=2.0.0)",
    "psycopg2-binary (>=2.9.0)",
    "pydantic (>=2.6.0)",
    "pytest (>=8.0.0)",
    "python-dotenv (>=1.0.0)",
    "alembic (>=1.13.0)",
    "bcrypt (<4.0.0)",
    "python-jose[cryptography] (>=3.3.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib (>=1.7.4)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[dependency-groups]
dev = [
    "ipdb (>=0.13.13,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
